refactor(lib): extract merge-mode + view building into merge module (#217)
Move the contiguous PluginMergeMode enum, decide_merge_mode, the atomic
view/merged tree builders (dispatch_plugin_merge, build_view_atomically,
atomic_replace_view_dir, ensure_absent, sibling_with_suffix), and the
conflict aggregation/printing (record_merge_result, print_merge_conflicts)
out of lib.rs into src/merge.rs.

Functions are pub(crate); PluginMergeMode is re-exported from the crate root
so rvpm::PluginMergeMode stays valid. plural_s stays in lib.rs (used widely)
and is now pub(crate) so merge.rs can call it.

Behaviour-preserving code motion. cargo fmt --check + clippy -D warnings +
test (805) + --doc all green.
